+++ packages/plugin-admin-api/src/Plugin.js
@@ -9,7 +9,9 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
+const cors = require("cors");
 const express = require("express");
+
 const {
   DEFAULT_BASE_PATH,
   SETTINGS,
@@ -35,6 +37,10 @@ const OPTION = {
   type: "string",
   default: DEFAULT_BASE_PATH,
 };
+
+const corsMiddleware = cors({
+  preflightContinue: false,
+});
 
 class Plugin {
   static get id() {
@@ -95,6 +101,7 @@ class Plugin {
 
   _initRouter() {
     this._router = express.Router();
+    this._router.use(corsMiddleware);
     this._router.use(ABOUT, this._aboutApi.router);
     this._router.use(SETTINGS, this._settingsApi.router); // TODO, add config route. deprecate settings
     this._router.use(ALERTS, this._alertsApi.router);